перенес генерацию в мок, поделал сколько успел остальные правки
